Recipe to return first N chars of a text field.
For data with text fields that has things like MODEL, SERIAL NO etc., using a transformer to extract the first N chars can make a big difference. First version, we have to hard code the 'n' value in the recipe. At some point, we should be able to determine that with heuristics, with some additional code.